Added a verification to the delete an update path operations for posts, to check if the user_id in the token matches the user who owns the posts

+++ app/routers/post.py
@@ -56,6 +56,10 @@ def delete_post(id: int, db: Session = Depends(get_db),
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f'post with id {id} does not exist')
 
+    if post_query.first().user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Not authorized to perform requested action")
+
     post_query.delete(synchronize_session=False)
     db.commit()
     return {'message: Post was succesfully deleted'}
@@ -66,9 +70,14 @@ def update_post(id: int, post: schemas.Post, db: Session = Depends(get_db),
                 current_user :int = Depends(oauth2.get_current_user)):
 
     post_query = db.query(models.Post).filter(models.Post.id == id)
+
     if post_query.first() == None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"post with id: {id} does not exist")
+
+    if post_query.first().user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Not authorized to perform requested action")
 
     post_query.update(post.dict(), synchronize_session=False)
     db.commit()
